+++ src/Corvus.Json.Cli.Core/GenerationDriverV5.cs
@@ -244,6 +244,10 @@ public static class GenerationDriverV5
         bool excludeNonNullDefaulted = optionalAsNullableMode.Equals(GeneratorConfig.OptionalAsNullable.EnumValues.NullOrUndefinedExceptNonNullDefaulted);
         bool optionalAsNullable = excludeNonNullDefaulted || optionalAsNullableMode.Equals(GeneratorConfig.OptionalAsNullable.EnumValues.NullOrUndefined);
 
+        GeneratorConfig.NativeEnums nativeEnumsMode = generatorConfig.NativeEnumsValue ?? GeneratorConfig.NativeEnums.DefaultInstance;
+        bool emitNativeStringEnums = nativeEnumsMode.Equals(GeneratorConfig.NativeEnums.EnumValues.All) || nativeEnumsMode.Equals(GeneratorConfig.NativeEnums.EnumValues.StringEnums);
+        bool emitNativeFlagsEnums = nativeEnumsMode.Equals(GeneratorConfig.NativeEnums.EnumValues.All) || nativeEnumsMode.Equals(GeneratorConfig.NativeEnums.EnumValues.FlagsObjects);
+
         return new CodeGeneration.CSharpLanguageProvider.Options(
             (string)generatorConfig.RootNamespace,
             namedTypes: namedTypes.Select(n => new CodeGeneration.CSharpLanguageProvider.NamedType(new((string)n.Reference), (string)n.DotnetTypeName, n.DotnetNamespace?.GetString(), GetAccessibility(n) ?? defaultAccessibility)).ToArray(),
@@ -261,7 +265,9 @@ public static class GenerationDriverV5
             buildParametersThreshold: generatorConfig.BuildParametersThreshold is Corvus.Json.JsonInteger buildParametersThreshold && !buildParametersThreshold.IsUndefined()
                 ? (int)buildParametersThreshold
                 : CodeGeneration.CSharpLanguageProvider.Options.DefaultBuildParametersThreshold,
-            formatModeOverrides: GetFormatModeOverrides(generatorConfig));
+            formatModeOverrides: GetFormatModeOverrides(generatorConfig),
+            emitNativeStringEnums: emitNativeStringEnums,
+            emitNativeFlagsEnums: emitNativeFlagsEnums);
     }
 
     private static IReadOnlyDictionary<string, FormatAssertionMode>? GetFormatModeOverrides(in GeneratorConfig generatorConfig)