+++ tests/Corvus.Text.Json.Tests.GeneratedModels.NativeEnums/NotFlagsSingle.cs
@@ -1,0 +1,6 @@
+namespace Corvus.Text.Json.Tests.GeneratedModels.NativeEnums.Draft202012;
+
+[JsonSchemaTypeGenerator("./not-flags-single-2020-12.json")]
+public readonly partial struct NotFlagsSingle
+{
+}